using separate table component
