+++ src/pages/pieceDocumentation/index.css
@@ -26,7 +26,7 @@
     background-color: transparent;
     border: 2px solid var(--ifm-color-primary-dark);
     border-radius: 50%; 
-    padding: 10px 20px;
+    padding: 10px 15px;
     text-decoration: none;
     color: var(--ifm-color-primary-dark); 
     font-size: 16px;
